+++ .github/scripts/test_docker_image.sh
@@ -59,8 +59,8 @@ print_json = false
 [http_config]
 server_url = "'$KMS_URL_HTTPS'"
 accept_invalid_certs = true
-ssl_client_pkcs12_path = "'$CLIENT_PKCS12_PATH'"
-ssl_client_pkcs12_password = "password"
+tls_client_pkcs12_path = "'$CLIENT_PKCS12_PATH'"
+tls_client_pkcs12_password = "password"
 ' | tee "$TLS_CONFIG"
 
 # Use cargo run to execute ckms from the workspace instead of installing